% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_events.R
\name{batch_distribution}
\alias{batch_distribution}
\title{Distribution Test}
\usage{
batch_distribution(data, value)
}
\arguments{
\item{data}{A \code{data.frame}}

\item{value}{A \code{numeric} vector}
}
\value{
A \code{list} with named distributions, e.g. Weibull,
GEV, Log-pearson type III, Pearson, Gumbel, Normal and Lognormal distributions.
}
\description{
This function uses multiple statistics package \link[fitdistrplus]{fitdistrplus-package},
\link[evd]{evd-package}, \link[smwrBase]{smwrBase-package} and the \link[stats]{stats} to generate
multiple theoretical and empirical univariate distributions, e.g. Weibull,
GEV, Log-pearson type III, Pearson, Gumbel, Normal and Lognormal distributions.
}
