% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_events.R
\name{batch_frequency}
\alias{batch_frequency}
\title{Batched Frequencies}
\usage{
batch_frequency(data, value)
}
\arguments{
\item{data}{A \code{data.frame} with \code{numeric} variable and date}

\item{value}{A \code{numeric} vector.}
}
\value{
A \code{data.frame} with return intervals and associated value for Weibull,
GEV, Log-pearson type III, Pearson, Gumbel, Normal and Lognormal distributions.
}
\description{
This function fits a variable (e.g., flow, snow, rain, etc) to a univariate distribution
function, e.g. Weibull,
GEV, Log-pearson type III, Pearson, Gumbel, Normal and Lognormal distributions.
It is up to the user to decide which distribution function to use and whether any pre-processing is necessary (outliers, mixed populations, etc).
}
