% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_StreamStats.R
\name{batch_RRE}
\alias{batch_RRE}
\title{Batch USGS Reginal Regression Estimates (RRE)}
\usage{
batch_RRE(state, wkID, group = NULL)
}
\arguments{
\item{state}{An abbreviated State \code{character}, e.g. "MT"}

\item{wkID}{A workspace ID generated in \link[wildlandhydRo]{batch_StreamStats}}
}
\value{
Returns a data.frame with associated regional regression estimates.
}
\description{
Provides the USGS regressions from a \link[wildlandhydRo]{batch_StreamStats} object or manually entered params.
Uses methods from \insertCite{ries2017streamstats}{wildlandhydRo} to generate RRE's.
}
\examples{
\dontrun{
# Bring in data

data <- tibble(state = "MT", wkID = "MT20201116184949291000")

rre_peak <- batch_RRE(state = data$state, wkID = data$wkID)

#### Or use batch_StreamStats object

data <- tibble(Lat = c(48.30602, 48.62952, 48.14946),
                 Lon = c(-115.54327, -114.75546, -116.05935),
                   Site = c("Granite Creek", "Libby Creek", "WF Blue Creek"))

three_sites <- batch_StreamStats(lon = data$Lon, lat = data$Lat, group = data$Site,
                                  crs = 4326)

rre_peak <- batch_RRE(state = three_sites$state, wkID = three_sites$wkID, group = three_sites$group)

}#' @importFrom Rdpack reprompt
}
\references{
{
\insertAllCited{}
}
}
