% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_StreamStats.R
\name{batch_StreamStats}
\alias{batch_StreamStats}
\title{Download Multiple Stream Stats Locations}
\usage{
batch_StreamStats(lon, lat, group = NULL, crs = 4326)
}
\arguments{
\item{lon}{A \code{numeric} vector of longitude values}

\item{lat}{A \code{numeric} vector of latitude values}

\item{group}{A vector to group by}

\item{crs}{A \code{numeric} crs value}
}
\value{
Returns an sf (simple feature) object with associated basin characteristics.
}
\description{
Takes longitude and latitude vectors and returns an sf object. Uses \link[AOI]{geocode_rev} to get
the state identifier and \link[streamstats]{computeChars} and \link[streamstats]{delineateWatershed} to generate basin
delineation(s) and characteristics,
which use methods from \insertCite{ries2017streamstats}{wildlandhydRo}
}
\examples{
\dontrun{
# Bring in data

data <- tibble(Lat = c(48.30602, 48.62952, 48.14946),
                 Lon = c(-115.54327, -114.75546, -116.05935),
                   Site = c("Granite Creek", "Louis Creek", "WF Blue Creek"))

three_sites <- batch_StreamStats(lon = data$Lon, lat = data$Lat, group = data$Site,
                                  crs = 4326)

}
}
\references{
{
\insertAllCited{}
}
}
