% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_StreamStats.R
\name{batch_StreamStatsCatchChars}
\alias{batch_StreamStatsCatchChars}
\title{Stream Stats Catchment Characteristics}
\usage{
batch_StreamStatsCatchChars(watersheds)
}
\arguments{
\item{watersheds}{A previously created \link[wildlandhydRo](batch_StreamStatsDelineate)}
}
\description{
Takes previously created batch_StreamStatsDelineate() and returns watershed boundary and catchment characteristics. Uses \link[streamstats]{computeChars}
to get basin characteristics,
which use methods from \insertCite{ries2017streamstats}{wildlandhydRo}
}
\examples{
\dontrun{
# Bring in data

data <- tibble(Lat = c(48.30602, 48.62952, 48.14946),
                 Lon = c(-115.54327, -114.75546, -116.05935),
                   Site = c("Granite Creek", "Louis Creek", "WF Blue Creek"))
data <- data \%>\% sf::st_as_sf(coords('Lon', 'Lat'))
three_sites <- batch_StreamStatsDelineate(data, group = 'Site',
                                  crs = 4326)
catch_chars <- batch_StreamStatsCatchChars(three_sites)

}
}
\references{
{
\insertAllCited{}
}
}
