% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_StreamStats.R
\name{batch_StreamStatsDelineate}
\alias{batch_StreamStatsDelineate}
\title{Download Multiple Stream Stats Locations}
\usage{
batch_StreamStatsDelineate(data, group, crs = 4326)
}
\arguments{
\item{data}{A \code{data.frame} with \code{lon,lat} variables. \code{optional}}

\item{group}{A vector to group by. \code{optional}}

\item{crs}{A \code{numeric} crs value}
}
\value{
Returns an sf (simple feature) object with associated basin characteristics.
}
\description{
Takes sf point object and returns watershed boundary (sf). Uses \link[AOI]{geocode_rev} to get
the state identifier and \link[streamstats]{computeChars} and \link[streamstats]{delineateWatershed} to generate basin
delineation(s) and characteristics,
which use methods from \insertCite{ries2017streamstats}{wildlandhydRo}
}
\examples{
\dontrun{
# Bring in data

data <- tibble(Lat = c(48.30602, 48.62952, 48.14946),
                 Lon = c(-115.54327, -114.75546, -116.05935),
                   Site = c("Granite Creek", "Louis Creek", "WF Blue Creek"))
data <- data \%>\% sf::st_as_sf(coords('Lon', 'Lat'))
three_sites <- batch_StreamStatsDelineate(data, group = 'Site',
                                  crs = 4326)

}
}
\references{
{
\insertAllCited{}
}
}
