% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{clean_USGSdv}
\alias{clean_USGSdv}
\title{Batch USGS daily values}
\usage{
clean_USGSdv(
  Sites,
  parameterCd = "00060",
  start_date = "",
  end_date = "",
  statCd = "00003"
)
}
\arguments{
\item{Sites}{}
}
\value{
A data.frame
}
\description{
Batch USGS daily values
}
\examples{


}
