% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_StreamStats.R
\name{culvert_size}
\alias{culvert_size}
\title{Culvert Size}
\usage{
culvert_size(x)
}
\arguments{
\item{x}{A vector of \code{numeric} flows}
}
\value{
A \code{data.frame}
}
\description{
This function let's you enter any numeric vector of flows and returns recommended culvert
sizes based on \insertCite{american1983handbook}{wildlandhydRo}.
}
\references{
{
\insertAllCited{}
}
}
