% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dl_ws}
\alias{dl_ws}
\title{Delineate watersheds}
\usage{
dl_ws(df)
}
\arguments{
\item{df}{A filtered data.frame}
}
\value{
hopefully a geojson feature collection. if not then either in AK or server is not responding correctly
}
\description{
Hijacked streamstats function delineateWatershed() to add a retry() call because
frequently the function will send a 'set vector 0/0' but really just needs to be ran again.
This is the work-around.
}
