% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_StreamStats.R
\name{download_StreamStats}
\alias{download_StreamStats}
\title{Download Multiple Stream Stats Locations

Looking at stream stats}
\usage{
download_StreamStats(lon, lat, state, group = NULL, crs = NULL)
}
\arguments{
\item{lon}{A numeric vector of longitude values}

\item{lat}{A numeric vector of latitude values}

\item{state}{An abbreviated State character, e.g. "MT"}

\item{group}{A variable to group by, e.g. "site"}

\item{crs}{A numeric crs value, e.g. 4326 (default)}

\item{data}{A data.frame with lat and lon variables}
}
\value{
Returns an sf (simple feature) object with associated basin characteristics.
}
\description{
Download Multiple Stream Stats Locations

Looking at stream stats
}
\examples{
\dontrun{
# Bring in data

data <- tibble(Lat = c(48.660, 48.652), Lon = c(-115.542, -115.517), Site = c("who knows", "oh yeah"))

two_sites <- download_StreamStats(lon = data$Lon, lat = data$Lat, state = "MT", group = data$Site, crs = 4326)

}
}
