% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_getbasin.R
\name{get_Basin}
\alias{get_Basin}
\title{Get Basin Boundary NLDI}
\usage{
get_Basin(sf_pt)
}
\arguments{
\item{sf_pt}{A sf point object or atomic list.}
}
\value{
An sf object.
}
\description{
Get Basin Boundary NLDI
}
