% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_getbasin.R
\name{get_BasinStats}
\alias{get_BasinStats}
\title{Get Basin Stats}
\usage{
get_BasinStats(data)
}
\arguments{
\item{data}{A previously created get_Basin() object or an sf object with comid column.}
}
\value{
An sf object with stats.
}
\description{
This function will get pre-loaded basin statistics (zonal stats)
for a provided comid. A comid can be generated in the get_Basin().
}
