% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_events.R
\name{get_exceedance}
\alias{get_exceedance}
\title{Get Exceedance Probability}
\usage{
get_exceedance(x, a = 0.375)
}
\arguments{
\item{x}{A numeric vector}

\item{a}{A numeric value for Gringoten formula}
}
\value{
A data.frame with exceedance probabilities and return periods.
}
\description{
Takes a numeric vector to find plotting position methods for exceedance probability.
}
\note{
Methods used are: California, Hazen, Weibull, Chegodayev, Blom and Gringoten.
}
