% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_events.R
\name{get_exceedence}
\alias{get_exceedence}
\title{Title}
\usage{
get_exceedence(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A data.frame with exceedance probabilities and return periods.
}
\description{
Title
}
