% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_StreamStats.R
\name{get_flow_basin}
\alias{get_flow_basin}
\title{Delineate watersheds}
\usage{
get_flow_basin(watersheds, st)
}
\arguments{
\item{watersheds}{A nested data.frame}

\item{st}{A vector with state abbreviations}
}
\value{
two tibbles. one with flow stats and the other with basin stats
}
\description{
Hijacked streamstats function computeChars() and writeGeoJSON to run a for loop getting the geojson polygon and associated
polygon stats
}
