% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_events.R
\name{get_RI}
\alias{get_RI}
\title{Get Reoccurrence Intervals}
\usage{
get_RI(freqDV, values, distr, RI = NULL)
}
\arguments{
\item{freqDV}{A previously create \link[wildlandhydRo]{batch_frequency} object.}

\item{values}{A \code{numeric} vector}

\item{distr}{Distribution from a batch_frequency object, see notes.}

\item{RI}{A \code{numeric} vector. default is \code{NULL}}
}
\value{
A tibble print out and ggplot graph.
}
\description{
Get Reoccurrence Intervals
}
\note{
The distributions that are acceptable: "Gumbel", "LogPearson", "GEV", "Normal", "Lognormal", "Pearson", "Weibull".
}
