% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_SNOTEL.R
\name{get_SNOTEL}
\alias{get_SNOTEL}
\title{Get SNOTEL locations}
\usage{
get_SNOTEL(AOI = NULL)
}
\arguments{
\item{AOI}{A sf object}
}
\value{
A \code{data.frame}
}
\description{
A function that locates the nearest SNOTEL station and returns a sf object. This function only works with
a sf object and is really handy when using
\link[AOI]{AOI-package} functions. This function is hijacked from \link[nhdplusTools]{get_NWIS}.
}
