% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_SNOTEL.R
\name{hourlySNOTEL}
\alias{hourlySNOTEL}
\title{Hourly SNOTEL}
\usage{
hourlySNOTEL(Sites = NULL, procDV = NULL, days = 7)
}
\value{
A data.frame with hourly SNOTEL metrics
}
\description{
This function gets hourly SNOTEL data from "https://wcc.sc.egov.usda.gov/reportGenerator/" website.
}
