% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_USGS_30day.R
\name{laura_DeCicco_fun}
\alias{laura_DeCicco_fun}
\title{30-day moving average}
\usage{
laura_DeCicco_fun(
  procDV,
  site = NULL,
  rolln = 30,
  startDate = "2010-01-01",
  endDate = "2015-01-01"
)
}
\arguments{
\item{procDV}{A previously created proc DV}

\item{site}{A USGS NWIS site}

\item{rolln}{Number of days in moving average}
}
\value{

}
\description{
30-day moving average
}
