% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_data}
\alias{meta_data}
\title{SNOTEL Meta Data}
\format{
A data frame with site info.
}
\usage{
meta_data
}
\description{
Meta data from snotelr::snotel_info() function
}
\keyword{datasets}
