% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_SNOTEL.R
\name{monthSNOTEL}
\alias{monthSNOTEL}
\title{Month-Only Stats (SNOTEL)}
\usage{
monthSNOTEL(procDV)
}
\arguments{
\item{wymDV}{A \link[wildlandhydRo]{proc_SNOTELdv} object.}
}
\value{
A data.frame with month only statistics for snow water equivalent and snow depth.
}
\description{
Month-Only Stats (SNOTEL)
}
\examples{


}
