% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{monthUSGS}
\alias{monthUSGS}
\title{Month-Only Stats (USGS)}
\usage{
monthUSGS(procDV)
}
\arguments{
\item{procDV}{A previously created \link[wildlandhydRo]{proc_USGSdv} object.}
}
\value{
A \code{data.frame}
}
\description{
This function uses the results of the \link[wildlandhydRo]{proc_USGSdv} object to
generate mean, maximum, median and standard deviation for month-only.
}
\examples{


}
