% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_events.R
\name{plot_allDist}
\alias{plot_allDist}
\title{Plot Distribution Plots Together}
\usage{
plot_allDist(procDist)
}
\arguments{
\item{procDist}{A previously created \link[wildlandhydRo]{batch_distribution} object}
}
\value{
A ggplot object. Can't use with \link[plotly]{ggplotly}
}
\description{
This function plots all of the  \link[wildlandhydRo]{wildlandhydRo-package}
plot_*Dist functions. Mostly for reporting but also good for a quick EDA.
}
