% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_events.R
\name{plot_densDist}
\alias{plot_densDist}
\title{Plot Density}
\usage{
plot_densDist(procDist, facet = TRUE)
}
\arguments{
\item{procDist}{A previously created \link[wildlandhydRo]{batch_distribution} object}

\item{facet}{Ability to facet by distribution. default \code{TRUE}}
}
\value{
A ggplot object. Can use with \link[plotly]{ggplotly}
}
\description{
This function plots a \link[wildlandhydRo]{batch_distribution} object using
 a probability density function.
}
