% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_events.R
\name{plot_reportDist}
\alias{plot_reportDist}
\title{Plot Distribution Report}
\usage{
plot_reportDist(report)
}
\arguments{
\item{report}{A previously created \link[wildlandhydRo]{reportDist} object}
}
\value{
A ggplot object. Can use with \link[plotly]{ggplotly}
}
\description{
This function plots \link[wildlandhydRo]{reportDist} object. it provides
the user graphs for each statistic.
}
