% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_SNOTEL.R
\name{plot_reportSNOTEL}
\alias{plot_reportSNOTEL}
\title{Plot SNOTEL Report}
\usage{
plot_reportSNOTEL(report, time = "daily", type = "per")
}
\arguments{
\item{report}{A previously created \link[wildlandhydRo]{batch_SNOTELdv} object}

\item{time}{A string ("month", "daily") indicating what time frame to plot.}

\item{type}{A string ("per", "now", "all") indicating what type of plot.}
}
\description{
Plot SNOTEL Report
}
