% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{plot_reportUSGS}
\alias{plot_reportUSGS}
\title{Plot USGS Report}
\usage{
plot_reportUSGS(report, time = "daily")
}
\arguments{
\item{report}{A previously created \link[wildlandhydRo]{reportUSGSdv} or \link[wildlandhydRo]{reportUSGSmv} object.}

\item{time}{A \code{character} vector indicating "daily" or "month".}
}
\value{
A ggplot object
}
\description{
This plots the results of the daily and monthly report from a \link[wildlandhydRo]{reportUSGSdv} or \link[wildlandhydRo]{reportUSGSmv} object.
This function will take more than one station but be cautious as it gets too muddled fast, i.e.
two is fine but three may be too much.
}
