% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{plot_USGSdrought}
\alias{plot_USGSdrought}
\title{USGS Drought Plot}
\usage{
plot_USGSdrought(
  procDV,
  site = NULL,
  rolln = 30,
  startDate = "2010-01-01",
  endDate = "2015-01-01",
  smooth.span = NULL
)
}
\arguments{
\item{procDV}{A previously created \link[wildlandhydRo]{batch_USGSdv} object.\code{optional}}

\item{site}{A USGS NWIS site. \code{optional}}

\item{rolln}{A \code{numeric} number of days in moving average}

\item{startDate}{A \code{character} indicating the start date.}

\item{endDate}{A \code{character} indicating the end date.}

\item{smooth.span}{A \code{numeric} value indicating the smoothing parameter for the \link[stats]{loess} function. If NULL (default) no smoothing will happen.}
}
\value{
A ggplot. Works with plotly::ggplotly.
}
\description{
This function is hijacked from a blog post by Laura DeCicco at \url{https://waterdata.usgs.gov/blog/moving-averages/}.
I changed it a little to include compatibility with \link[wildlandhydRo]{batch_USGSdv} and also the ability to
change the rolling average, e.g. \code{rolln}. For more details visit the blog. Also, this function is expensive so it
will take about 5 to 15 seconds to produce plot depending on period of record.
}
