% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{plot_USGSfdc}
\alias{plot_USGSfdc}
\title{USGS Flow Duration Curve}
\usage{
plot_USGSfdc(
  procDV,
  site = NULL,
  startDate = "2010-01-01",
  endDate = "2015-01-01",
  seasons = FALSE,
  span_season = NULL
)
}
\arguments{
\item{procDV}{A previously created \link[wildlandhydRo]{batch_USGSdv} object.\code{optional}}

\item{site}{A USGS NWIS site. \code{optional}}

\item{startDate}{A \code{character} indicating the start date.}

\item{endDate}{A \code{character} indicating the end date.}

\item{seasons}{A \code{logical} whether to split by seasons.}

\item{span_season}{A \code{character} indicating a season, e.g. '4-1':'6-30'.}
}
\value{
A ggplot. Works with plotly::ggplotly.
}
\description{
This function generates Flow Duration Curves.
}
