% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_SNOTEL.R
\name{proc_SNOTELdv}
\alias{proc_SNOTELdv}
\title{Process SNOTEL daily values}
\usage{
proc_SNOTELdv(Sites)
}
\value{
A data.frame with SNOTEL metrics.
}
\description{
This function gets daily snotel data from "https://wcc.sc.egov.usda.gov/reportGenerator/" website. This
function is similar to \link[snotelr]{snotel_download} except it includes english units and adds snow depth as a variable.
}
