% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{proc_USGSdv}
\alias{proc_USGSdv}
\title{Process USGS daily values}
\usage{
proc_USGSdv(
  sites,
  parameterCd = "00060",
  start_date = "",
  end_date = "",
  statCd = "00003"
)
}
\arguments{
\item{sites}{A vector of USGS NWIS sites}
}
\value{
A \code{data.frame} with daily mean flow and added meta-data.
}
\description{
This function is basically a wrapper around \link[dataRetrieval]{readNWISdv} but includes
added variables like water year, lat/lon, station name, altitude and tidied dates.
}
\examples{


}
