% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_events.R
\name{reportDist}
\alias{reportDist}
\title{Summary Statistic Distribution Report}
\usage{
reportDist(procDist)
}
\arguments{
\item{procDist}{A previously created \link[wildlandhydRo]{batch_distribution} object}
}
\value{
A \code{data.frame}
}
\description{
This function uses a \link[wildlandhydRo]{batch_distribution} to tidy a
\link[fitdistrplus]{gofstat} function. Good for choosing which distribution to use. For more details
see the \link[fitdistrplus]{fitdistrplus-package} on how to use the output.
}
