% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{reportUSGSdv}
\alias{reportUSGSdv}
\title{USGS Report Daily}
\usage{
reportUSGSdv(procDV, sites = NULL)
}
\arguments{
\item{procDV}{A previously created \link[wildlandhydRo]{batch_USGSdv} object.}

\item{sites}{A \code{character} USGS NWIS site.}
}
\value{

}
\description{
This function uses the \link[dataRetrieval]{readNWISstat} to gather daily
statistics like quantiles.
}
