% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_evd.R
\name{rfrechet}
\alias{rfrechet}
\title{Hi-jacked \link[evd]{evd-package} functions for working internally in package}
\usage{
rfrechet(n, loc = 0, scale = 1, shape = 1)
}
\value{

}
\description{
Hi-jacked \link[evd]{evd-package} functions for working internally in package
}
