% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snotel_functions.R
\name{snotel_metric_custom}
\alias{snotel_metric_custom}
\title{snotel custom functions}
\usage{
snotel_metric_custom(df)
}
\arguments{
\item{data.frame}{A data.frame}
}
\value{
The daily precip in snotel
}
\description{
Takes any data.frame
}
