% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{wyUSGS}
\alias{wyUSGS}
\title{Water Year Stats (USGS)}
\usage{
wyUSGS(procDV)
}
\arguments{
\item{procDV}{A previously created \link[wildlandhydRo]{proc_USGSdv} object.}
}
\value{

}
\description{
This function uses the results of the \link[wildlandhydRo]{proc_USGSdv} object to
generate mean, maximum, median and standard deviation per water year. It also includes peaks from
\link[dataRetrieval]{readNWISpeak} and normalization
by \code{log(Flow)/log(drainage area)} and \code{log(Flow)/log(all time mean flow)}.
}
