% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_events.R
\name{batch_frequency}
\alias{batch_frequency}
\title{Batched Frequencies}
\usage{
batch_frequency(data, value)
}
\arguments{
\item{data}{A \code{data.frame} with \code{numeric} variable and date}

\item{value}{A \code{numeric} vector. length > 10}
}
\value{
A \code{data.frame} with return intervals and associated value for Weibull,
GEV, Log-pearson type III, Pearson, Gumbel, Normal and Lognormal distributions.
}
\description{
Batched Frequencies
}
