% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_StreamStats.R
\name{batch_RRE}
\alias{batch_RRE}
\title{Batch USGS Regional Regression Estimates (RRE)}
\usage{
batch_RRE(data = NULL, state, wkID, group)
}
\arguments{
\item{data}{A \code{data.frame} with \code{state and wkID} variables}

\item{state}{An abbreviated State \code{character}, e.g. "MT"}

\item{wkID}{A workspace ID generated in \link[wildlandhydRo]{batch_StreamStats}}

\item{group}{A vector to group by. \code{optional}}
}
\value{
Returns a data.frame with associated regional regression estimates.
}
\description{
Provides the USGS regressions from a \link[wildlandhydRo]{batch_StreamStats} object or manually entered params.
Uses methods from \insertCite{ries2017streamstats}{wildlandhydRo} to generate RRE's.
}
\examples{
\dontrun{
# Bring in data

data <- tibble(state = "MT", wkID = "MT20201116184949291000")

rre_peak <- batch_RRE(state = data$state, wkID = data$wkID)

#### Or use batch_StreamStats object

data <- tibble(Lat = c(48.3060, 48.6293, 48.1494),
                 Lon = c(-115.5432, -114.7554, -116.0593),
                   Site = c("Granite Creek", "Louis Creek", "WF Blue Creek"))

three_sites <- batch_StreamStats(lon = data$Lon, lat = data$Lat, group = data$Site,
                                  crs = 4326)

rre_peak <- batch_RRE(state = three_sites$state, wkID = three_sites$wkID, group = three_sites$group)

}
}
\references{
{
\insertAllCited{}
}
}
