% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_StreamStats.R
\name{batch_RRE}
\alias{batch_RRE}
\title{Batch USGS Regional Regression Estimates (RRE)}
\usage{
batch_RRE(data)
}
\arguments{
\item{data}{A previously created \link[wildlandhydro]{batch_StreamStats} object.}
}
\value{
Returns a data.frame with associated regional regression estimates.
}
\description{
Provides the USGS regressions from a \link[wildlandhydRo]{batch_StreamStats} object or manually entered params.
Uses methods from \insertCite{ries2017streamstats}{wildlandhydRo} to generate RRE's.
}
\examples{
\dontrun{
# Bring in data

#### use batch_StreamStats object

data <- tibble(Lat = c(48.30602, 48.62952, 48.14946),
                 Lon = c(-115.54327, -114.75546, -116.05935),
                   Site = c("Granite Creek", "Louis Creek", "WF Blue Creek"))

data <- data \%>\% sf::st_as_sf(coords('Lon', 'Lat'))

three_sites <- batch_StreamStats(data, group = 'Site',
                                  crs = 4326)

rre_peak <- batch_RRE(three_sites)

}
}
\references{
{
\insertAllCited{}
}
}
