% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_USGS_30day.R
\name{laura_DeCicco_fun}
\alias{laura_DeCicco_fun}
\title{30-day moving average}
\usage{
laura_DeCicco_fun(
  procDV,
  site = NULL,
  rolln = 30,
  startDate = "2010-01-01",
  endDate = "2015-01-01",
  smooth.span = NULL
)
}
\arguments{
\item{procDV}{A previously created proc DV}

\item{site}{A USGS NWIS site}

\item{rolln}{Number of days in moving average}

\item{startDate}{A character date, e.g. '2000-01-01'}

\item{endDate}{A character date, e.g. '2002-12-25'}
}
\description{
30-day moving average
}
